Extract tests to test suit
