+++ arhelk-lexer.cabal
@@ -24,8 +24,6 @@ library
                      , text
                      , text-show
                      , parsec
-                     , QuickCheck
-                     , quickcheck-instances
 
   default-language:    Haskell2010
   default-extensions:  OverloadedStrings
@@ -61,11 +59,16 @@ test-suite arhelk-lexer-test
                      , arhelk-lexer
                      , test-framework
                      , test-framework-quickcheck2
+                     , QuickCheck
                      , quickcheck-instances
-
+                     , text
+                     , parsec
   ghc-options:         -threaded -rtsopts -with-rtsopts=-N
   default-language:    Haskell2010
-
+  other-modules:       Test.Arhelk.Lexer.Grammar
+                       Test.Arhelk.Lexer.Token
+  default-extensions:  OverloadedStrings
+  
 source-repository head
   type:     git
   location: https://github.com/Teaspot-Studio/arhelk-lexer.git
